Module clif_json::serde_clif_json

Structs

SerBlock
Serializable version of Cranelift IR Blocks.

SerDataFlowGraph
Serializable Data Flow Graph.

SerFunction
Serializable Function type, including name, signature, global values, and data flow graph.

SerInst
Serializable version of Cranelift IR instructions.

SerObj
Must have SerObj for deserialization, contains all of the functions from inside the file to be serialized. Files have one SerObj each, with all SerFunctions contained inside that SerObj.

SerSignature
Serializable signature including function parameters and returns.

Enums

SerInstData
Serializable version of the original Cranelift IR

Functions

get_inst_data
Convert Cranelift IR instructions to JSON format.

populate_blocks
Serialize all parts of the Cranelift Block data structure, this includes name, parameters, and instructions.

populate_inst

populate_params
Translating Block parameters into serializable parameters.